contractContext.js was splitted into multiple files and moved to src/context/contract/
