+++ src/context/contract/types.js
@@ -1,0 +1,18 @@
+export const START = 'START';
+export const FAIL = 'FAIL';
+
+export const GET_DISPLAY_CORGIS = 'GET_DISPLAY_CORGIS';
+export const GET_CORGISLIST_SUCCESS = 'GET_CORGISLIST_SUCCESS';
+export const GET_CORGI_SUCCESS = 'GET_CORGI_SUCCESS';
+
+export const CREATE_START = 'CREATE_START';
+export const CREATE_CORGI_SUCCESS = 'CREATE_CORGI_SUCCESS';
+
+export const TRANSFER_START = 'TRANSFER_START';
+export const TRANSFER_CORGI_SUCCESS = 'TRANSFER_CORGI_SUCCESS';
+
+export const DELETE_START = 'DELETE_START';
+export const DELETE_CORGI_SUCCESS = 'DELETE_CORGI_SUCCESS';
+
+export const CLEAR = 'CLEAR';
+export const CLEAR_CREATED_CORGI_STATE = 'CLEAR_CREATED_CORGI_STATE';
